app/src/main/java/com/expense/ui/main/MainActivity.kt
package com.expense.ui.main

import android.os.Bundle
import androidx.activity.ComponentActivity
import androidx.activity.compose.setContent
import androidx.activity.viewModels
import androidx.compose.foundation.ExperimentalFoundationApi
import androidx.compose.runtime.Composable
import androidx.compose.ui.tooling.preview.Preview
import com.expense.ui.theme.ExpenseManagementTheme
import dagger.hilt.android.AndroidEntryPoint

@ExperimentalFoundationApi
@AndroidEntryPoint
class MainActivity : ComponentActivity() {

    val viewModel by viewModels<ExpenseViewModel>()
    override fun onCreate(savedInstanceState: Bundle?) {
        super.onCreate(savedInstanceState)
        setContent {
            MainView(viewModel)
        }
    }

}


@Preview(showBackground = true)
@Composable
fun DefaultPreview() {
    ExpenseManagementTheme {
        MainView(viewModel)
    }
}

